Theresa's Uploaded Files
I uploaded my Income Statement, Coach Schedule and Unpaid Debt pages. I also altered the sign in/up pages so it has a section for the treasurer to go. Right now because the treasurer homepage is not uploaded, I just have it redirect to one of my pages for now to make sure it could load.